+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed update getting stuck at 100% or close, fix applied to install part too
 
+### Changed
+
+- Updated classic mode default window size to better match the 1440p background images
+
 ## [3.15.2] - 29.07.2025
 
 ### Fixed